+++ lua/obsidian/config.lua
@@ -454,7 +454,7 @@ end
 ---@field folder string|obsidian.Path|?
 ---@field date_format string|?
 ---@field time_format string|?
----@field substitutions table<string, function|string>|?
+---@field substitutions table<string, (fun(ctx: obsidian.TemplateContext):string)|(fun(): string)|string>|?
 config.TemplateOpts = {}
 
 --- Get defaults.
@@ -465,6 +465,9 @@ config.TemplateOpts.default = function()
     folder = nil,
     date_format = nil,
     time_format = nil,
+    -- A map for custom variables, the key should be the variable and the value a function.
+    -- Functions are called with obsidian.TemplateContext objects as their sole parameter.
+    -- See: https://github.com/obsidian-nvim/obsidian.nvim/wiki/Template#substitutions
     substitutions = {},
   }
 end